+++ internal/gdi/font/font_js.go
@@ -1,9 +1,7 @@
-package gdi
+package font
 
 // -------------------------------------------------------------------------------------
 
 var fontFindPaths = []string{
-	"/system/fonts/",
+	"/Library/Fonts",
 }
-
-// -------------------------------------------------------------------------------------
