+++ internal/gdi/font/font_linux.go
@@ -1,11 +1,9 @@
-package gdi
+package font
 
 // -------------------------------------------------------------------------------------
 
 var fontFindPaths = []string{
-	"/Library/Fonts",
-	"/System/Library/Fonts",
-	"/System/Library/Fonts/Supplemental",
+	"/system/fonts/",
 }
 
 // -------------------------------------------------------------------------------------
