+++ internal/gdi/font/font_windows.go
@@ -1,0 +1,9 @@
+package font
+
+// -------------------------------------------------------------------------------------
+
+var fontFindPaths = []string{
+	"C:\\windows\\fonts",
+}
+
+// -------------------------------------------------------------------------------------
