--- internal/gdi/font_js.go
@@ -1,7 +1,0 @@
-package gdi
-
-// -------------------------------------------------------------------------------------
-
-var fontFindPaths = []string{
-	"/Library/Fonts",
-}
